# api/openapi.yaml
openapi: "3.0.3"
info:
  title: "LICEU API"
  description: >
    API de analise e cruzamento de dados publicos brasileiros.
    Integra dados de multiplas fontes governamentais: TSE, TCU,
    Portal da Transparencia, PNCP, Congresso Nacional (Camara e Senado),
    IBGE e SICONFI.
  version: "0.1.0"
  contact:
    name: "LICEU"
    url: "https://github.com/danyele/laceu"
servers:
  - url: "http://localhost:8080"
    description: "Desenvolvimento local"
...
paths:
...
  /deputados/{id}/completo:
    get:
      tags: [Deputados]
      summary: "Busca detalhes completos de um deputado"
      operationId: buscarDetalhesDeputado
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: "Detalhes do deputado"
          content:
            application/json:
              schema:
                type: object
        "400":
          description: "ID invalido"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Erro"
...
  /estado/{uf}/candidatos:
    get:
      tags: [Estado]
      summary: "Busca candidatos de um estado"
      operationId: buscarCandidatosEstado
      parameters:
        - name: uf
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: "Candidatos do estado"
          content:
            application/json:
              schema:
                type: object
  /estado/{uf}/deputados:
    get:
      tags: [Estado]
      summary: "Busca deputados de um estado"
      operationId: buscarDeputadosEstado
      parameters:
        - name: uf
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: "Deputados do estado"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DadosWrapper"
...
  /portal-transparencia/servidores/funcoes-e-cargos:
    get:
      tags: [PortalTransparencia]
      summary: "Busca funcoes e cargos"
      operationId: buscarFuncoesECargos
      parameters:
        - name: pagina
          in: query
          schema:
            type: integer
            default: 1
        - name: codigoFuncaoCargo
          in: query
          schema:
            type: string
        - name: descricaoFuncaoCargo
          in: query
          schema:
            type: string
      responses:
        "200":
          description: "Funcoes e cargos"
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
...
  /portal-transparencia/despesas/por-funcional-programatica:
    get:
      tags: [PortalTransparencia]
      summary: "Busca despesas por funcional programatica"
      operationId: buscarPorFuncionalProgramatica
      parameters:
        - name: ano
          in: query
          schema:
            type: integer
        - name: pagina
          in: query
          schema:
            type: integer
            default: 1
        - name: funcao
          in: query
          schema:
            type: string
        - name: subfuncao
          in: query
          schema:
            type: string
        - name: programa
          in: query
          schema:
            type: string
        - name: acao
          in: query
          schema:
            type: string
      responses:
        "200":
          description: "Despesas por funcional programatica"
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
...
  /portal-transparencia/despesas/funcional-programatica/listar:
    get:
      tags: [PortalTransparencia]
      summary: "Lista funcional programatica completa"
      operationId: listarFuncionalProgramatica
      parameters:
        - name: ano
          in: query
          schema:
            type: integer
        - name: pagina
          in: query
          schema:
            type: integer
            default: 1
      responses:
        "200":
          description: "Funcional programatica"
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
  /portal-transparencia/despesas/documentos:
    get:
      tags: [PortalTransparencia]
      summary: "Busca documentos de despesa"
      operationId: buscarDocumentos
      parameters:
        - name: pagina
          in: query
          schema:
            type: integer
            default: 1
        - name: dataEmissao
          in: query
          schema:
            type: string
        - name: fase
          in: query
          schema:
            type: string
        - name: unidadeGestora
          in: query
          schema:
            type: string
      responses:
        "200":
          description: "Documentos"
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
...
  /portal-transparencia/despesas/documentos-relacionados:
    get:
      tags: [PortalTransparencia]
      summary: "Busca documentos relacionados"
      operationId: buscarDocumentosRelacionados
      parameters:
        - name: codigoDocumento
          in: query
          schema:
            type: string
        - name: fase
          in: query
          schema:
            type: string
      responses:
        "200":
          description: "Documentos relacionados"
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
...
  /portal-transparencia/despesas/plano-orcamentario:
    get:
      tags: [PortalTransparencia]
      summary: "Busca despesas por plano orcamentario"
      operationId: buscarPlanoOrcamentario
      parameters:
        - name: ano
          in: query
          schema:
            type: integer
        - name: pagina
          in: query
          schema:
            type: integer
            default: 1
        - name: codPlanoOrcamentario
          in: query
          schema:
            type: string
        - name: descPlanoOrcamentario
          in: query
          schema:
            type: string
      responses:
        "200":
          description: "Plano orcamentario"
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
...
components:
  schemas:
    Erro:
      type: object
      properties:
        erro:
          type: string
          description: "Mensagem de erro"
      required:
        - erro
    DadosWrapper:
      type: object
      properties:
        dados:
          description: "Dados da resposta (tipo varia por endpoint)"
...